requirements.txt
flask
requests
python-dotenv
pytesseract
pillow
PyPDF2
python-docx
Flask-Limiter
